requirements.txt
azure-storage-blob
openai
langchain
langchain-community
langchain-openai
tiktoken
python-dotenv
streamlit
streamlit-extras
streamlit-cookies-manager
openpyxl
tabulate
pandas
numpy
frontend
tools
PyMuPDF
python-docx
graphviz